requirements.txt
# YouTube API & Data Collection
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
isodate==0.6.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# OpenAI for Embeddings
openai==1.12.0

# Document Processing
PyMuPDF==1.23.26
python-docx==1.1.0

# Data Processing & ML
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

# Storage
pyarrow==14.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For progress bars
tqdm==4.66.1

# Sentence Transformers - CHỈ GIỮ MỘT PHIÊN BẢN
sentence-transformers>=5.0.0
torch>=2.0.0
huggingface-hub>=0.19.0